Moved MRC programs from INTEGER*2 to INTEGER*4
git-svn-id: svn+ssh://svncina.ethz.ch/repos/2dx/trunk@1737 cb738334-da44-0410-bf9e-9ace271e2f5b

+++ kernel/mrc/source/2dx_edmtz.for
@@ -79,7 +79,7 @@ C
       INTEGER IHI(MAXPTS),IKI(MAXPTS),ILI(MAXPTS)
       REAL FPIN(MAXPTS),PHIN(MAXPTS),SIGFPIN(MAXPTS),FOMIN(MAXPTS)
 C
-      INTEGER*2 IBEGIN(IRNGE,IRNGE),IFINISH(IRNGE,IRNGE),IP1,IP2
+      INTEGER*4 IBEGIN(IRNGE,IRNGE),IFINISH(IRNGE,IRNGE),IP1,IP2
 C
 C---- NLPRGI  =  number of input labels
       DATA NLPRGI,LSPRGI/7,'H','K','L','FP','SIGFP','PHS',
@@ -399,12 +399,12 @@ C  Arms length routine to call ASYM. All matrices for manipulating
 C   H,K's are in this subroutine instead of .MAIN (or other)
 C
       LOGICAL LSPEC
-      INTEGER*2 IP1,IP2
+      INTEGER*4 IP1,IP2
 C
-      INTEGER*2 ISPEC(5,17)
+      INTEGER*4 ISPEC(5,17)
       DATA ISPEC/7*0,1,3*0,1,4*0,1,4*0,1,3*0,3*1,2*0,3*1,0,-1,3*1,0,1,
      A 3*1,4*0,1,0,0,4*1,0,5*1,8*0,1,0,1,1,5*0,1,4*0,1,1,0/
-      INTEGER*2 IGO(8,17)
+      INTEGER*4 IGO(8,17)
       DATA IGO/8*5,2*4,2*5,2*4,2*5,
      A 4,5,4,5,4,5,4,5,  4,5,4,5,4,5,4,5,  4,5,4,5,4,5,4,5,
      B 2,4,2,5,2,4,2,5,  2,4,2,5,2,4,2,5,  2,4,2,5,2,4,2,5,
@@ -417,14 +417,14 @@ C       THE FIRST ELEMENT OF EACH IS PASSED TO SET,ASYM FOR LATER USE.
 C       THE SAME IS DONE FOR IGO WHICH CONTROLS PROGRAM FLOW IN SET,ASYM
 C       AND FOR ISPEC WHICH INDICATES SPECIAL REFLECTIONS.
 C
-      INTEGER*2 IMAT(5,17)
+      INTEGER*4 IMAT(5,17)
       DATA IMAT/ 1,1,1,1,1,    1,2,1,1,1,    1,3,1,1,1,
      A           1,4,1,1,1,    1,3,1,1,1,    1,2,1,3,1,
      B           1,2,1,4,1,    1,2,1,6,1,    1,2,1,3,1,
      C           1,2,7,5,1,    1,8,1,2,3,    1,8,1,9,6,
      D           1,10,11,12,1, 1,8,1,10,11,  1,9,1,10,11,
      E           1,2,10,5,11,  1,8,9,10,11/
-      INTEGER*2 MAT(8,12)
+      INTEGER*4 MAT(8,12)
       DATA MAT/   -1,0,0,-1,-1,0,0,-1,      1,0,0,1,-1,0,0,-1,
      A            1,0,0,-1,1,0,0,-1,        1,0,0,-1,1,0,180,-1,
      B            0,1,-1,0,1,0,0,1,        1,0,0,-1,1,180,180,-1,
@@ -517,8 +517,8 @@ C
 C*************************************************************************
       SUBROUTINE ASYM(IH,IK,Z,IP1,IP2,SPEC,IPTEST,WSTAR,
      1  A1,A2,A3,A4,A5,IGO,ISPEC,LREV)
-      INTEGER*2 A1(8),A2(8),A3(8),A4(8),A5(8),IGO(8),ISPEC(5)
-      INTEGER*2 IP1,IP2
+      INTEGER*4 A1(8),A2(8),A3(8),A4(8),A5(8),IGO(8),ISPEC(5)
+      INTEGER*4 IP1,IP2
       LOGICAL SPEC,LREV
 C
 C      WRITE(6,904)A1,A2,A3,A4,A5,IH,IK,Z,IP1,IP2,SPEC,IPTEST,WSTAR
@@ -629,7 +629,7 @@ C                             0      0     0      0  IA(8)
 C           FOR ALL CASES.
 C
 C
-      INTEGER*2 IA(8),IP1,IP2
+      INTEGER*4 IA(8),IP1,IP2
 C      WRITE(6,900)IA,IH,IK,Z,IP1,IP2
       IH1=IA(1)*IH+IA(2)*IK
       IK=IA(3)*IH+IA(4)*IK
